+++ tstore/tslong/pandas.py
@@ -2,7 +2,6 @@
 
 from typing import TYPE_CHECKING
 
-import pandas as pd
 import pyarrow as pa
 
 from tstore.archive.io import (
@@ -14,8 +13,6 @@ from tstore.archive.io import (
 from tstore.archive.metadata.writers import write_tstore_metadata
 from tstore.archive.partitions import add_partitioning_columns, check_partitioning
 from tstore.archive.ts.writers.pyarrow import write_partitioned_dataset
-from tstore.tsdf.ts_class import TS
-from tstore.tsdf.tsarray import TSArray
 from tstore.tslong.pyarrow import TSLongPyArrow
 from tstore.tslong.tslong import TSLong
 
@@ -166,62 +163,10 @@ class TSLongPandas(TSLong):
 
     def to_tsdf(self) -> "TSDFPandas":
         """Convert the wrapper into a TSDF object."""
-        from tstore.tsdf.pandas import TSDFPandas
-
-        tstore_ids = self._get_tstore_ids()
-        ts_arrays = self._get_ts_arrays()
-        pd_series = {ts_variable: pd.Series(ts_array, index=tstore_ids) for ts_variable, ts_array in ts_arrays.items()}
-        static_values = self._get_static_values()
-        data = {**pd_series, **static_values, self._tstore_id_var: tstore_ids}
-
-        df = pd.DataFrame(data)
-        return TSDFPandas(
-            df,
-            id_var=self._tstore_id_var,
-        )
-
-    def _get_tstore_ids(self) -> list:
-        """Retrieve the list of tstore ids."""
-        return sorted(self._obj[self._tstore_id_var].unique())
-
-    def _get_ts_arrays(self) -> dict[str, TSArray]:
-        """Create a dictionary of TSArrays."""
-        return {ts_variable: self._get_ts_array(variables) for ts_variable, variables in self._tstore_ts_vars.items()}
-
-    def _get_ts_array(self, variables: list[str]) -> TSArray:
-        """Create a TSArray for a set of variables."""
-        tstore_ids = self._get_tstore_ids()
-        ts_list = [self._get_ts(tstore_id, variables) for tstore_id in tstore_ids]
-        return TSArray(ts_list)
-
-    def _get_ts(self, tstore_id: str, variables: list[str]) -> TS:
-        """Create a TS object for a given tstore_id and a set of variables."""
-        df = self._obj
-        df = df[df[self._tstore_id_var] == tstore_id]
-        df = df[[self._tstore_time_var, *variables]]
-        return TS(df)
-
-    def _get_static_values(self) -> dict[str, list]:
-        """Retrieve the static values."""
-        df = self._obj
-        tstore_ids = self._get_tstore_ids()
-        static_values = {}
-
-        for static_var in self._tstore_static_vars:
-            var_static_values = []
-            for tstore_id in tstore_ids:
-                values = df[df[self._tstore_id_var] == tstore_id][static_var].unique()
-
-                if len(values) > 1:
-                    raise ValueError(
-                        f"Static variables should be unique per tstore_id. Found {list(values)} for {static_var}.",
-                    )
-
-                var_static_values.append(values[0])
-
-            static_values[static_var] = var_static_values
-
-        return static_values
+        dask_tslong = self.change_backend(new_backend="dask")
+        dask_tsdf = dask_tslong.to_tsdf()
+        tsdf = dask_tsdf.change_backend(new_backend="pandas")
+        return tsdf
 
     def to_tswide(self) -> "TSWidePandas":
         """Convert the wrapper into a TSWide object."""
